refactor: move KeyringVersion internally into keyring-snap-bridge
