+++ packages/keyring-api/src/index.ts
@@ -5,4 +5,3 @@ export * from './eth';
 export type * from './contexts';
 export * from './rpc';
 export * from './events';
-export * from './versions';
